% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runEpiEstimApp}
\alias{runEpiEstimApp}
\title{Runs EpiEstimApp}
\usage{
runEpiEstimApp(...)
}
\arguments{
\item{...}{Additional arguments (passed to shiny::runApp())}
}
\description{
This function runs EpiEstimApp
}
\examples{
\dontrun{
runEpiEstimApp()
}
}
\keyword{run}
