% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runEpiEstimApp}
\alias{runEpiEstimApp}
\title{Runs EpiEstimApp}
\usage{
runEpiEstimApp(port = 3000)
}
\arguments{
\item{port}{The port to run on. Defaults to 3000.}
}
\description{
This function runs EpiEstimApp
}
\examples{
runEpiEstimApp()
}
\keyword{run}
